+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v0.1.223
+
 - Inject the `EntityNormalizer` into `AbstractAuthorizationService` for testing to avoid null references
 - TestClient: offer a method to reset the request caches of a given `AbstractAuthorizationService` service